- quaternion improvements
